Extract Z3950Provider to its own file
